renamed package api.bukkit.config.custom.handler to api.bukkit.config.custom.handle
